requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
PyJWT==2.4.0
requests==2.28.1
python-dotenv==0.20.0